Add authentication
Enable Warden for authentication
Create an "API token" authentication strategy which uses a simple
username & token passed in the HTTP headers to identify & validate a
user. This is *not* cryptographically secure (that would require some
form of request signing) but the method can be extended later.
Add a User model, DB migration & test user fixture.
Use authentication in the controller.