Add a PUT endpoint for Organizations
Allow Organizations to be modified. The owner_email & users list can be
modified; modifying the Users list allows Users to be added & removed from
an Organization.
Remove the "add user to listed organizations" code from the Users endpoint in
favour of the "Create user, add user to Organization" model.
Re-order the User & Organization end-points so that the different verbs for
the same endpoints are together.