Add checks for Super Admins
Add the is_super_admin? helper method; any user that is a member of the
'admins' group is considered a Super Admin.
If a user is a Super Admin, promote the organization that's being authorized
to 'admins'. This means we can have "Read-only Super Admins" (A user who is a
member of the 'admins' organization where the Userpermissions are
admin=False, write=False, read=True): such a user would have read-only access
to every organization.
Rename the admins group from 'cyclid' to 'admins'