Re-factor some of the Action & Plugin code
Rename Action to Step
Re-factor the Stage definition so that Stages have Steps, which in turn
define an action: the action name then maps directly onto a plugin name.
Modify the way in which Action plugins are initialized so that they can
be serialised/unseralised easily.
Add the preare method to the Action plugin so that the transport & context
information can be passed in after the plugin instance has been initialized.