Make the SSH transport actually...work
As it was, the SSH transport looked very nice and neat but didn't actually
work in a useful fashion; commands were async so you couldn't rely on a
command completing before the next one was run, nor was a useful & reliable
way to retrieve the return code of a command. So...
Don't use a shell for the SSH channel.
Create a new channel for each command (I.e. each time exec is called)
Build a command string (export variables, cd to the cwd, run the command)
with build_command.
execute the command string on the channel & run the session loop until
it's completed, then retrieve the return code.
Pass the path to the transport during exec so that it can cd to it as
part of the command string.