Abstract JobRecord & LogBuffer from Job::Runner
Remote job runners will not be able to access AR models such as JobRecord
directly, so we need to abstract that access away via. a Notifier class;
the Notifier also proxies writes to the LogBuffer (as the LogBuffer also
wants to write to the log in the JobRecord). A Notifier for a remote
Runner can pump the updates onto a message queue, to be picked up and
translated into updates the JobRecord at the API application end. A local
Notifier can just write to JobRecord directly.