Add Sidekiq
Add a Sidekiq worker for the Local dispatcher which executes the job Runner,
and invoke it asyncronously rather than running the Runner inside the
Sinatra worker context.
Add Sidekiq to the Gemfile, Gaurdfile and Rakefile tasks for Redis & Sidekiq.
Configure a Rack map for both Cyclid ('/') & the Sidekiq dashboard
('/sidekiq'); this is fine for testing for now but not suitable for
production.