Add the concept of Notifier callbacks
It's useful to be able to hook into the Notifier methods so that external code
(E.g. plugins) can become aware of the job status; to this end, add a callback
object which is serialised and passed to the Sidekiq worker, which then creates
a new Callback instance and pings the methods on it when an event occures
(status changes, job is completed, data is written to the log).