Pass useful HTTP headers to methods
Sinatra provides a 'headers' helper but that only returns additional headers
and ignores X- headers; the Rack env has all of the raw headers but that might
expose data (and is ugly). Instead, process the Rack env to extract headers
that are a) safe & b) useful, and then pass those on to the method.