Split set_config & update_config
There's no need (or desire) for plugins to have to deal with Cyclid models
(E.g. Organization & PluginConfig), and no need for each plugin to re-implement
it's own code to deal with them. So make set_config deal with getting/creating
the current configuration, and saving the updated configuration.
Move the actual logic for validating & updating the configuration into the
update_config method, which takes the current & new configs and returns an
updated (merged) configuration that can be serialised back into the plugin
config.
Consistently use strings for keys.
Serialize & unserialize the "config" column in & out of the plugin config.