Clean up
Clean up a bunch of the original code; mostly make use of the URLs provided in
the request body rather than building our own artisinal versions. Also clear
up some of the confusion about what are URI instances and what are just URLs
as a String instance; anything with the _url suffix is a URI instance and
anything without is a String.
Inject the source details (URL, branch & OAuth token) into the job so that the
Job Runner retrieves the branch for the PR automatically.