Fix API plugins
I wasn't thinking my cunning plan all the way through; as it was, you'd only
ever be able to register one set of callback methods as they were being
stored in a class variable.
Store the methods in an instance variable and then register them as helpers
externally, when the controller is registered. This works.