Add some sort of SSH transport tests
Add some test for the SSH transport plugin.
Testing with Net::SSH mocks is incredibly complicated and fragile; there is no
real Mocks for it, only Net::SSH::Test which seems to be ill concieved and not
very useful for any serious testing. So, these tests largely do no work.