Add encryption data to the organization
Add the rsa_private_key, rsa_public_key columns for storing an RSA key pair
(in DER format).
Add the salt column for storing a simple salt to be used with other encryption
schemes (E.g. AES)
Create an RSA key-pair and salt when an organization is created.
Ensure that the encryption data is sanitized from the organization records, but
put the public key into the organization document (DER encoded with Base64) so
that organization members can retrieve it.
Update the tests to reflect the change; make some fuzzy-matches as they won't
know what the generated public key data is, and add a test to ensure the RSA
public key is properly returned to an organization member.
Update db-loader to create organizations with valid RSA keys & salts.