Make SSH keys work with Transports
The SSH transport can't find the SSH private key on its own when it's not
running as an interactive user, so we need to pull the path to the key
from the Mist config and then pass it in.
Extend the BuildHost::connect_info method to return the key data.
The Mist client also needs the list of servers from the Mist config to
create the Pool.