Start to clean job context data
Fix the way that the environment data is added to the job context.
Rename some of the context names to be more obvious E.g. job_name to match
job_id
Add the organization name to the job & inject that into the context data.
Allow the plugin get_config method to accept the organization name as a string
and use it to retrieve the organization object itself, to avoid exposing the
Organization model directly to plugins.