Allow context to be attached to jobs
Add a 'context' argument to the job creator helper and allow callers to pass
aribitary data when creating a new job, which is then stored as context data
in the JobView.
Use the initial job context data at the start of the job in the job Runner.
Modify the tests to pass the additional context argument or context data where
required.
Use the interpolated URL argument in the Slack notification plugin.
Fix a few Rubocops.