Make the app accessable from the Gem
Now that the code is under app/ we can add a lib/ directory with a shim that
can be require'd E.g. by config.ru
Update all of the paths that expand lib/ to expand app/ instead.
Update the Guardfile to use the common shim under lib/ instead of the old
Sidekiq specific init.rb