Fix stage insertion
The logic for handling when a stage is referenced in an on_success or
on_failure handler but not explitely defined in the job was all
horribly wrong, leading to it over-writing the proper stage definitions
with minimal objects with much of the data missing.
The has_stage? method is a more complex (and correct) way to check if a stage
exists in the sequence.