Make the API plugins more flexible
The quad of verbs with a single endpoint is just not flexible enough; this
provides a way for advanced plugins to add arbitary endpoints (within their
own namespace) to the API, too.
Use the new functionality to DRY up the default HTTP methods.
Make a start to add the OAuth hooks to the Github plugin, including loading a
server-wide plugin config which can contain the Github client ID etc.