Generate & pass a "state" for all OAuth requests
Github can take/return an opaque "state" parameter which can be used to ensure
that the callback request came from Github (and not somebody spoofing a
callback). For that to work we need to generate a unique hash, pass it to
the OAuth request and check it on the callback.