Improved plugin configurations
Add the GET /organization/:org/config endpoint to return a list of plugins
that have configurations.
Add the has_config? method to plugins.
Add the 'password' schema type for sensitive configuration data.
Add the 'link-relative' schema type for links: these are not really
"configuration" but they are related.
Add the OAuth request link to the Github plugin schema.