Fix more problems with PR's from Forks
Use the correct status URL
Correct the helpers & logic around the PR "head" and "base"; head is always
the *source* and base is always the *target*; always clone from the head and
push updates to the base.
Improve the logic of inserting the Pull Request repository into the sources;
the comment was prescient of the issue, so ensure that the head repository URL
has been normalized into a Cyclid source definition, and then normalize the
entire set of sources so that we *over-write* any existing source definition
for the base repository, with the head repository. This means, for example,
that we don't try to clone github.com/fork/foo on top of
github.com/example/foo.
Updates tests to match these changes.