Provide an Evaluator class that can parse & evaluate simple statements in the form "lvalue operator rvalue", and returns either the result (only_if) or inverted result (not_if) of the evaluated expression. The evaluator supports both String and Number comparisions. Strings are in the form "'lvalue' operator 'rvalue'" I.e. enclosed within single quotes. Numbers can be integers, floats or percentages. Operators supported are ==, != and === for Strings, and ==/eq, !=/ne, </lt, >/gt, <=/lte & >=/gte for Numbers.
Ensure that only_if/not_if expressions are stored in the StageView.
Evaluate any only_if or not_if expression that's defined, and either
run or skip the current stage depending on the outcome; if no
modifier is defined, always run the stage. If the stage is skipped due
to the modifer, proceed as if the Stage has suceeded.

Add String::** to provide a "safe" version of the % (interpolation) method;
** handles missing keys by inserting a nil value, instead of throwing a
KeyError.
Use ** instead of % to provide safe job-context interpolation everywhere.
Rename the Hash::interpolate and Array::interpolate methods to Hash::% and
Array::% to match String; eventually these both fall back to String::** to
both are "safe" by default.