Re-order sequence logic
The stage for the on_success & on_failure handler should be selected *prior*
to ensuring it's inserted into the sequence, otherwise we run the risk that
we'll select an on_failure handler from the options block that isn't
subsequently in the Job definition. We may as well clean up the on_success
code to have a similar logical flow while we're here, which also has the nice
side-effect of making it slightly more readable.