Add the fail_if modifier
The fail_if modifier is evaluated after the stage has run, and will force the
stage to be failed if the condition evaulates as true: this could be used to
E.g. check if test coverage is acceptable after the coverage plugin has run,
and fail the stage if the result is not acceptable.