Improve job validation
Lint every job with the standard linter and fail early if it does not pass
without errors.
Only create the JobRecord once the job has been successfully linted, and set
the Job name as soon as the JobRecord is created.
Raise specific exceptions rather than returning an HTTP error directly from
the helper, and handle those exceptions in the caller.
Provide more information to Github if a Pull Request or Push event fails due
to an error with the job definition.
Remove some simple job validation in the POST /organizations/:name/job
endpoint.