Improve sudo use
Only run a command under sudo when the caller specifically requires it via.
the 'sudo' argument to Transport::exec
Improve sudo invocation by invoking a shell to run the command under.
Set the 'sudo' argument where we need it internally.
Change the Transport#exec method to accept additional arguments, including
'path', as a hash. This is a breaking change for any plugins which use
Transports or derive from the Transport plugin.