Fix a major bug in #release
The buildhost hash uses symbol keys, so we should use :name to extract the
instance name to destroy it. Before this fix the GCP API #get method will be
called with `nil` for the name, and return the first instance it finds: this
works in simple tests (I.e. one instance only) but obviously fails horribly
if we have more than one instance (E.g. concurrent jobs)